Iteration 5.

+++ Week_6/week6-fepp-starter/frontend/src/hooks/useSignup.js
@@ -4,26 +4,33 @@ import { useNavigate } from "react-router-dom";
 const useSignup = (setIsAuthenticated) => {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
+    const [confirmationPassword, setConfirmationPassword] = useState('');
     const navigate = useNavigate();
 
     const handleSignup = async () => {
         try {
-            const response = await fetch("/api/users/signup", {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ email, password }),
-            });
+            if (password === confirmationPassword) {
+              const response = await fetch("/api/users/signup", {
+                  method: "POST",
+                  headers: {
+                      "Content-Type": "application/json",
+                  },
+                  body: JSON.stringify({ email, password }),
+              });
+  
+              if (response.ok) {
+                  const user = await response.json();
+                  localStorage.setItem("user", JSON.stringify(user));
+                  console.log("User signed up successfully!");
+                  setIsAuthenticated(true);
+                  navigate("/");
+              } else {
+                  console.error("Signup failed", response);
+              }
 
-            if (response.ok) {
-                const user = await response.json();
-                localStorage.setItem("user", JSON.stringify(user));
-                console.log("User signed up successfully!");
-                setIsAuthenticated(true);
-                navigate("/");
             } else {
-                console.error("Signup failed", response);
+              console.log("Re-enter confirmation passwor");
+              
             }
         } catch (error) {
             console.error("Error during signup:", error);
@@ -34,6 +41,8 @@ const useSignup = (setIsAuthenticated) => {
         setEmail,
         password,
         setPassword,
+        confirmationPassword,
+        setConfirmationPassword,
         handleSignup
     };
 };
